# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP proxy с управлением доступом по IP (whitelist/blacklist/graylist), rate limiting (RPS/RPM/RPH/RPD/bandwidth/соединения) и мониторингом в реальном времени.",
        "title": "Proxy API",
        "contact": {
            "name": "API Support",
            "email": "artiomk00o@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/ip/check": {
            "get": {
                "description": "Определяет, в каком списке находится IP-адрес (black/gray/white/none).\nПорядок проверки: blacklist → graylist → whitelist → none.\nIP можно передать через query-параметр, тело запроса (JSON-строка) или он будет взят из RemoteAddr.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Проверить принадлежность IP к спискам",
                "parameters": [
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Проверяемый IP-адрес",
                        "name": "ip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат проверки",
                        "schema": {
                            "$ref": "#/definitions/handlers.CheckIpResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный IP-адрес",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CheckIpResponse": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "list": {
                    "type": "string",
                    "enum": [
                        "white",
                        "black",
                        "gray",
                        "none"
                    ],
                    "example": "white"
                },
                "message": {
                    "type": "string",
                    "example": "доступ разрешён"
                }
            }
        },
...
    }
}